+++ pganonymize/utils.py
@@ -16,8 +16,8 @@ from pgcopy import CopyManager
 from psycopg2.sql import SQL, Composed, Identifier
 from tqdm import trange
 
-from pganonymizer.constants import DEFAULT_CHUNK_SIZE, DEFAULT_PRIMARY_KEY
-from pganonymizer.providers import provider_registry
+from pganonymize.constants import DEFAULT_CHUNK_SIZE, DEFAULT_PRIMARY_KEY
+from pganonymize.providers import provider_registry
 
 
 def anonymize_tables(connection, definitions, verbose=False, dry_run=False):